+++ .gitignore
@@ -1,0 +1,5 @@
+.env
+/server/.env
+/server/venv/
+/server/__pycache__/
+
