--- server/test.py
@@ -1,129 +1,0 @@
-from flask import Flask, jsonify, request
-from flask_cors import CORS
-import google.generativeai as genai
-from dotenv import load_dotenv
-import os
-
-load_dotenv()
-
-app = Flask(__name__)
-
-CORS(app)
-
-GOOGLE_API_KEY=os.getenv('GOOGLE_API_KEY')
-
-genai.configure(api_key=GOOGLE_API_KEY)
-model = genai.GenerativeModel('gemini-pro')
-
-
-@app.route('/')
-def index():
-    return "Hello 👋."
-
-
-@app.route('/tree', methods=["POST", "GET"])
-def tree():
-    if request.method == 'POST':
-        data = request.get_json()
-        query = data.get('query')
-        response = model.generate_content('''I will give you a topic and you have to generate an explanation of the topic in points in hierarchical tree structure and respond with JSON structure as follows:
-        {
-            "name": "Java",
-            "children": [
-                {
-                "name": "Development Environment",
-                "children": [
-                    {
-                    "name": "Java Source Code",
-                    "value": ".java files",
-                    "description": "Human-readable code written with Java syntax."
-                    },
-                    {
-                    "name": "Java Development Kit (JDK)",
-                    "children": [
-                        {
-                        "name": "Compiler",
-                        "value": "translates to bytecode",
-                        "description": "Transforms Java source code into bytecode instructions understood by the JVM."
-                        },
-                        {
-                        "name": "Java Class Library (JCL)",
-                        "value": "predefined classes and functions",
-                        "description": "Provides a collection of reusable code for common functionalities."
-                        }
-                    ]
-                    }
-                ]
-                },
-                {
-                "name": "Execution",
-                "children": [
-                    {
-                    "name": "Java Runtime Environment (JRE)",
-                    "children": [
-                        {
-                        "name": "Java Virtual Machine (JVM)",
-                        "value": "executes bytecode",
-                        "description": "Software program that interprets and executes bytecode instructions."
-                        },
-                        {
-                        "name": "Class Loader",
-                        "value": "loads bytecode into memory",
-                        "description": "Loads .class files containing bytecode into JVM memory for execution."
-                        }
-                    ]
-                    },
-                    {
-                    "name": "Bytecode",
-                    "value": ".class files (platform-independent)",
-                    "description": "Machine-independent instructions generated by the compiler, executable on any system with JVM."
-                    },
-                    {
-                    "name": "Just-In-Time (JIT) Compilation (optional)",
-                    "value": "improves performance by translating bytecode to machine code",
-                    "description": "Technique that translates frequently used bytecode sections into native machine code for faster execution."
-                    }
-                ]
-                },
-                {
-                "name": "Key Features",
-                "children": [
-                    {
-                    "name": "Object-Oriented Programming",
-                    "value": "uses objects and classes",
-                    "description": "Programs are structured around objects that encapsulate data and behavior."
-                    },
-                    {
-                    "name": "Platform Independent (write once, run anywhere)",
-                    "value": "bytecode runs on any system with JVM",
-                    "description": "Java code can be compiled once and run on any platform with a JVM installed."
-                    },
-                    {
-                    "name": "Garbage Collection",
-                    "value": "automatic memory management",
-                    "description": "JVM automatically reclaims memory from unused objects, simplifying memory management for developers."
-                    }
-                ]
-                }
-            ]
-            }
-            Topic is: ''' + query)
-        
-        # print(response.text)
-        return jsonify({'success': True, 'data': response.text})
-        # return temp 
-
-
-@app.route('/interview', methods=["POST", "GET"])
-def interview():
-    if request.method == 'POST':
-        data = request.get_json()
-        if data.get('from') == 'client':
-            return "Success"
-        elif data.get('from') == 'gradio':
-            print(data)
-            return "Success"
-
-
-if __name__ == '__main__':
-    app.run()
